+++ server.js
@@ -10,6 +10,8 @@ var cors = require('cors')
 var config = require('./config'); // get our config file
 
 var schoolHandlers = require('./handlers')
+var comunaHandlers = require('./handlersComunas')
+var regionHandlers = require('./handlersRegion')
 
 var mongoDb = process.env.MONGO_DATABASE || config.database;
 console.log(mongoDb)
@@ -38,7 +40,26 @@ app.get('/education', function (req, res) {
   schoolHandlers.education(req,res);
 });
 
+app.get('/comuna', function (req, res) {
+  comunaHandlers.getAllComunas(req,res);
+});
 
+app.get('/comuna/:comuna', function (req, res) {
+  comunaHandlers.getOneComuna(req,res);
+});
+
+
+app.get('/region', function (req, res) {
+  regionHandlers.getAllRegion(req,res);
+});
+
+app.get('/region/:region', function (req, res) {
+  regionHandlers.getOneRegion(req,res);
+});
+
+app.get('/pais', function (req, res) {
+  regionHandlers.getPais(req,res);
+});
 
 app.listen(port, function () {
   console.log(`Listening on port ${port}!`);