+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+## [1.0.6] - 2017-04-23
+### Added
+- Retrieve data for Pais, Region, Comuna in a single year or all years (by default latest year)
+- Results include all metrics for matricula
+
+
 ## [1.0.4] - 2017-04-03
 ### Added
 - Additional metric for number of students in education levels